refactor footer component
